LayoutTests/fast/css-grid-layout/grid-item-start-before-get-set-expected.txt

Test that setting and getting grid-column-start and grid-row-start works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-column-start and grid-row-start set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row') is 'auto'
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-column') is "first / auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-column-start') is "first"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-row') is "last / auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-row-start') is "last"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-column') is "firstArea / auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-column-start') is "firstArea"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-row') is "thirdArea / auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-row-start') is "thirdArea"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-row-end') is "auto"

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto'

Test getting and setting grid-column-start and grid-row-start through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "nav / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "last / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "thirdArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "thirdArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "secondArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "nonExistentArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "nonExistentArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "secondArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "secondArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "nonExistentArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "nonExistentArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"

Test setting grid-start and grid-before to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "last / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "a / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "a"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "a / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "gridArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "gridArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "last / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "gridArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "gridArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"

Test setting grid-start and grid-before to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "a / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "first / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "first / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "first"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "auto / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "5"

Test setting grid-column-start and grid-row-start back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
